feat: Move to jiff datetime library.
Chrono is now soft-deprecated. Much of the Rust ecosystem is
switching to jiff. These changes move to Jiff for all but the
interface with polars. Polars is still using chrono, and we
cannot drop the chrono dependency until polars does.

There are some minor integer type changes. The biggest change
is the removal of the `Date` enum in favour of just using
`jiff::civil::DateTime` directly. However, this changes the
roundtrip deserlisation/serialisation behaviour losing the
information on whether the user provided a simple date or a
date-time.

The conversion tests currently fail because they are not updated
to reflect this change.

+++ pywr-core/src/parameters/mod.rs
@@ -890,7 +890,7 @@ pub enum ParameterBuildError {
         source: MetricF64Error,
     },
     #[error("Could not compute day of the year; invalid date: day: {day}, month: {day}")]
-    InvalidDayOfYear { day: u32, month: u32 },
+    InvalidDayOfYear { day: i8, month: i8 },
     #[error("Parameter is configured without a valid calculation phase: {detail}")]
     NoCalculationPhase { detail: String },
 }
